requirements.txt
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1649077682618/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1670263926556/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
boto3==1.26.133
botocore==1.29.133
certifi==2023.5.7
charset-normalizer==3.1.0
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1679481329611/work
debugpy @ file:///Users/runner/miniforge3/conda-bld/debugpy_1680755682866/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
distro==1.8.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
idna==3.4
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1682176699712/work
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1679336661730/work
ipython @ file:///Users/runner/miniforge3/conda-bld/ipython_1682783380619/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
jmespath==1.0.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1681432441054/work
jupyter_core @ file:///Users/runner/miniforge3/conda-bld/jupyter_core_1678994370610/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
numpy==1.24.3
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas==2.0.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1682644429438/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1677600924538/work
psutil @ file:///Users/runner/miniforge3/conda-bld/psutil_1681775196473/work
psycopg2-binary==2.9.6
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1681904169130/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2023.3
PyYAML==6.0
pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1679317086968/work
requests==2.30.0
s3transfer==0.6.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
SQLAlchemy==2.0.13
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
tabula-py==2.7.0
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1681817807031/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1678559861143/work
tzdata==2023.3
urllib3==1.26.15
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1677313463193/work
